Tweaks the logic for initializing env name to support capacity parameters

+++ event/inter.go
@@ -4,6 +4,6 @@ import "context"
 
 type Eventer interface {
 	GetName() string
-	GetConcurrence() int32
+	GetConcurrence() int64
 	Handle(ctx context.Context, data []byte)
 }
